Move liftoff.abz animations to a class
Encapsulates the code for reading liftoff.abz animations in a C++ class.
This localizes resource management and automatically frees resources
when leaving scope. The implementation is stylistically rough to plainly
show how code is being moved rather than beginning with a complete
rewrite.
